Remove 'modified' from all SQL update statements
Instead of unconditionally updating the modified date, the new codebase
makes an intelligent check of before and after data to determine if a
row *actually* changed. So this column is now under code control and not
db control.

There was also an instance of last_synced in here too. Same thing: that's
determined by logic and not unconditional as part of a generic UPDATE
statement.