Add setLastSynced logic, both setting and honoring
This properly tracks the last time we synced (or even attempted to sync)
a project or task. And then it also checks this time and refuses to make
another sync attempt if a hardcoded value of 120 seconds hasn't passed.

This is the core of this update: I was getting 502 and 503 errors and
wondering if I was hitting API velocity checks. The whole idea behind
this was to not bother to make an API call if we *just* did it.