Upgrade app to use todoist api v9
About a month or so ago, the v8 API was pulled. The v9 upgrade was less
than trivial: nearly every table changed (column names, data types, and
sometimes columns themselves were added or removed). Although I built
from a strong base, this was in some ways a do-over.

* .env.dist upgraded to use the v9 endpoint
* Sync.php
** Improved error handling
** The second pass of the two-pass implementation: complete rewrite
** Removed all the helper methods for the original second pass
* Todoist.php
** New method to get archive (/archive endpoint new with v9)

The change to the second pass is a game changer. I no longer have to
carefully time when I create and complete tasks, with respect to my
personal/offline backup. I can create and close a dozen tasks before
running the backup, and it will still find them. This is because a
local copy of a task is no longer required for historical updates. I
can ask todoist itself.

This has also cut down the number of API calls per run from 100s (and
growing without limit) to just 25 or so (per project, not per item).

This commit completes the code update. The sql changes are coming next.