Update wrapper scripts
* manual-sync: improved messaging, support for TTL, log scans
* resume-manual-sync: secondary script, in case of a failed run

The main purpose of the latter is to not backup the database and to
reuse an existing log. Otherwise, it's basically the same thing.

Note: the TTL support was necessary for v8 and is now just a vestigial
appendage for v9. I may go back and remove all of that, for clarity.